Added the main program interface and initialization routine.
This program is intended to be run in a separate directory, and will
create a working subdirectory to store its protected files. This will
include things like settings and the databases used. The main program
is envisioned as a primarily interactive loop, but could implement a
script-only implementation as well.