Some basic functionality in place for new database scheme.
Working on trying to implement the basics of records first, including
the ability to add, remove, change, etc. Underlying shelve database mapped
out and add, remove, and list are currently functional albeit in a very
basic state. Next step is to finish basic functionality for records and
begin to work on files as well.