Finished hacking out initial database modification code.
Basic operations exist to add, remove, list and update records. Next step
is to refactor this code to reduce redundancy, e.g. by adding in a separate
loop function that accumulates a series of arguments or key,value pairs to
use in the actual database records functions. Following this, and perhaps
an update of the prompt utilites (for e.g. making the errormsg standard for
certain prompt types or subclasses that always display the same message), can
finally start on the file side of the database.