Initial pass at record modification complete.
A very basic framework now exists to work with record objects within the
database. This includes basic functions such as adding, changing, and
removing either whole records or their attributes. As it is now, this is
driven largely by user input, and is somewhat verbose/redundant. In the
future, this could be improved. Additionally, more informative prompt
information could be provided, for instance by making better use of
the ability to check records and their attributes.

Next step now is to begin to work on integrating the file side of the
database, such that files can be added and removed together with records.
Ideally, file paths should be stored as part of their corresponding
record's attributes in a standard way, and adding/removing records should
consequently add/remove the corresponding files as well.