Improved addition of files to records.
Added ability to add additional files in the add_attribute_loop by specifying
an attribute that falls under a set of reserved "for file only" keywords. Next,
need to add similar interface for removal and changing, and then also to add
support for adding multiple files from directory(ies).

Once this is all done, can focus on other aspects, such as improving the display
of database data to the end user.