Continuing with initial setup of searches.
First goal is to be able to run a single BLAST search using one or more
queries against one or more databases. Some things to consider moving
forward is to improve the setup scheme such that several parameters are
entered by the user, including:
    -type of search (e.g. BLAST, HMMer)
    -type of databases (e.g. protein, nucleotide, etc.)
    -ability to change parameters of the search itself
    -ability to keep/discard output files
At the very least, the first two need to be implemented before test
searches can be started.