Most basic BLAST search is almost within reach.
Can almost get to the first milestone of blasting one query into one
database. The issue is that BLAST does not seem to recognize the query
when provided as a sequence held in a single variable. Although BLAST
supposedly recognizes piped input, i.e. piping the output of "cat <file>"
into the BLAST program, it appears that it requires the context of a
"file-like" input. Therefore, might have to resort to using tmpfiles
or some other workaround.