Can BLAST from stdin now.
Eventually want to give users the option to save each inidividual query file as
a separate collection of files, or to simply run each separate search from the
relevant object as stdin without having to create separate files. The interface
is still very premature, but managed to get a BLAST search to run by only piping
the description and sequence of a query object into the BLAST command using
subprocess.