BLAST searches can be conducted, stored, and persist as objects.
Continuing towards eventual goal of making searches easy to conduct and
re-analyze at a later time. Currenlty, can perform a BLAST search and
parse the output, storing data regarding the output as well as the parsed
output itself in a result object. Now need to focus on analyzing search
results using user input, and conducting more complex searches such as
reverse BLAST and HMMer searches.

Some points have come up, which may require more detailed changes at
a later time:

-Wanted to use the new BioPython SearchIO module to parse the output, but
had to fall back on the parser found in Bio.Blast instead because the new
module uses lambda functions, which cannot be pickled, and hence could not
be stored in the shelve-style database. Might be able to still use this
in the future, but might have to populate a separate Object from the parsed
output first, as this can be written using only pickle-able objects.

-All of the various DBs (pickle files/shelve dbs) use the same core interface,
which should be centralized into a superclass and then extended as required.

-Currently, searches themselves and parsing are separate functions. However,
due to the nature of the implementation, Result objects are currently created
during the search so that they can be populated with relevant information.
This is not necessarily an issue, but it should be considered moving forward
when other objects/functions try to access the ResultDB - if the parsing
has not been carried out yet, then obviously the call will fail.