Continuing with search summarizing code.
Realized that a possible issue exists for the case with reverse searches. In the
forward search, it is designed such that each forward search consists of only one
query, rather than having multiple such searches in a single output file. This is
done so that the 'read' rather than 'parse' function of Bio.Blast can be used (the
latter uses a generator and hence cannot be pickled while the former uses a simple
list and can be pickled). However, reverse searches necessarily necessitate multiple
searches tied to the same original forward query (for multiple forward hits). Might
need to get a bit more complicated in how forward and reverse searches are linked,
potentially through yet another object or attribute, and this could be called on
later to make forward/reverse searches both picklable.