Paused before further development due to need for testing.
Need to begin to work out all the issues with the current implementation of
summarizing searches/results, which will be laborious. Previous noted issue
regarding multiple reverse search results may not be a large issue after all,
but this remains to be seen; still have not decided the best way to link
forward and reverse searches necessarily.