Continuing with search implementation.
Reverse searches now work, though have yet to attempt parsing. Discovered
an issue that needs to be sorted out. For some reason, parse() after
reverse search works for a few results and then begins to yield a SystemError
"Negative size passed to PyBytes_FromStringAndSize" for all subsequent
attempts to add parsed BLAST result information to records. This is a huge
issue and may force implementation differences if cannot figure out how to
fix. Obvious response is to simply force keeping output files and keep a
pointer to each file name for later parsing using BioPython, but this is not
ideal as it forces users to keep large xml files.