Confirmed that storing BLAST results in shelves will not work.
Decided to test another example of forward/reverse searching to determine
whether previously noted issue was a one-off or not. Does seem like storing
large objects as attributes of shelve entries for persistence will not work.
Problem is not well-described online, but appears to be an issue with the
length of the object's bytes surpassing that supported by __iter__ (i.e. it
sends a negative value as overflow which is not iterable). Therefore, should
enforce keeping output files and parsing on the fly using BioPython. This
also means it may no longer be necessary to split up results into individual
query files prior to reverse searches, though this remains to be seen.