In order to test summarizing forward/reverse results, need redundant accessions.
Beginning to work on redundant accession implementation for forward/reverse
searches and working out possible homologues. Idea is to search query against
the parent file and then do one of two things with the output file:

1) User manually enters equivalent accessions using a 'curses' or other
similar interface when presented with the output file.

2) One or more functions can be employed that attempt to discern which
hits are redundant. Probably will be based on clustering/ranked evalue
differences but this remains to be seen.