Starting to work through manual redundant accession choice input.
Idea is to have an interface that allows users to see all of the potential
redundant accessions (i.e. self vs. self BLAST output) and easily choose
which ones to actually use as redundant accessions. Current idea is to
perform the BLAST, parse the output, and display it through curses. Then,
can work through a way to choose based on this output. Optimally, could
do something with a split window display and the ability to choose between
the two to move/remove choices, but unsure of how to implement this at
the moment.

As an aside, curses in general might be a better choice to controlling
the code interactively than simply asking for raw input everytime. This
may cause some compatability issues, but if we can settle for not caring
about non-UNIX platforms it should be fine.