Working on implementing redundant accs as a part of parsing results.
First pass at adding in redundant accession data for determining whether or
not the hit is a homologue or not. Essentially, just need to check each hit
to be either the original query or a hit in the list of redundant accessions.

Next step will be to test all of the current code and move on to expanding
options for getting/summarizing results. Once the entire interface works,
also need to consider several things:

1) Changing from command line prompts to a full GUI interface (using either
urwid, tkinter, or both)

2) Restructuring to use a different database system, or at least cleaning
up current database implementation to subclass one or more generic supers
and remove redundant code

3) Adding additional search functionality (e.g. HMMer, HHSearch). Ideally,
should end up with an interface that allows extension for each new
program that an end-user might want to implement?

4) Add test code

5) Add logging