Working on summarizing homologues based on forward/reverse searches.
Initial thought regarding identifying matching pairs is slightly incorrect. The
first step is to match the forward result to the reverse result (based on the
comparison of fwd_result.query.identity and rev_result.query.original_query) and
then to compare each fwd_hit id to the identity of the rev_result to ensure that
the proper comparison is made. This splits up the comparison into two parts, one
that deals with gross identity comparison and then a more fine-grain part that
determines whether the hit is actually positive or not via evalue criteria.
Currently working on the latter part, but realized that the loop has a small
error in the implementation of "next_hits" and cutoffs. Need to go back and look
at this in more detail.

After this is finished, need to work on summarizing just one search result based
on evalue criteria, as the current implementation is also likely to change.

When this is done, can either move on to outputting human-readable summaries of
search results (spreadsheets, figures, etc) or can go back and start shoring up/
refactoring code. Will likely need to do this sooner than later so it might be a
logical point (before the volume gets too substantial).