Continuing to work through summarizing results.
Core logic to loop through negative results mapped out and appears to work.
Need to go through E value cutoffs more closely, as well as to determine
what '0.0' means from a perspective of comparing values.

Once this is done need to work on actually adding positive hits to the
database.