Starting to work on re-vamping old code and implementing GUI.
Changed access to settings file so that it does not require passing a ref to
'goat_dir' between functions/instances (should do something similar for the
DB as well). Need to consider how this might become an issue if the user decides
to move these internals around themselves, but for now it suffices.

Working on settings interface. Basic GUI pop-up window works and displays
settings along with their current values. Need to work on fleshing this out now.
Some ideas include:

-only updating the settings file if the value actually changed from the default
-having an 'apply' button that does not close the window, a 'submit' button that
changes the settings and closes, and a 'close' button that just closes