Now working through a GUI for database-related operations.
Plan is to have a tree view widget that allows for browsing of the database
and also with functionality to update individual records. For now, just
working on the display functionality.

As an aside, may need to investigate module imports in more depth; import of
main 'goat' module seems to yield different attribute dicts depending on
where the call occurs and causes global variables wrapping method calls in
the main module to fail. It seems that wrapping these in functions instead
to defer reference to application code fixes the issue for now, but this is
likely indicative of a more serious issue that needs to be looked into.