Initial database gui viewer finished.
Very basic GUI interface implements a two-panel system, with a tree view to the
database object and a display panel to the right for information about individual
elements of the database. Currently, the main issue is that long filepaths are
hidden instead of being wrapped. Minwidth of the label is specified in character
widths whereas the wraplength is in pixels, so cannot simply wrap to the min length.
Need to look into trying to configure the wraplength dynamically based on the size
of the bounding box for the widget.

Next task is to add functionality for adding/removing/modifying records and their
associated files. Need to carefully consider each operation, such that maximum
flexibility is afforded to the user.