Continuing to work towards new ZODB backend implementation for records.
Now have a basic framework flushed out for interacting with the ZODB database and
taking into account atomicity of transactions, etc. Need to change the GUI code to
deal with new classes and then to start working on additional GUI classes to add/
remove/change various attributes of the records.

Also, need to consider at what stage exceptions might be raised and react
accordingly. At bare minimum, if an operation fails the user should be notified
with some description of the problem. Either way, this should not hang the
application or cause it to quit.