Updated settings form to not flash with updates.
Initial implementation involved simply re-building the Toplevel window with additions/
deletions of settings. Work on database display led to the realization that this
behaviour is not ideal and that window updates should be forced without re-draws. As
a test, managed to make the current settings implementation follow this ideal.

Main change is that the settings form now keeps references to all EntryRow objects;
Tkinter does not have a built-in mechanism to easily access instances of child
widgets so the best way is to manually manage them by extending classes and storing
them in an appropriate data structure.

Plan now is to carry this over to the database GUI panel, such that adding new
attributes or file information causes updates to the main panel. The idea is that
users should be able to see all the information that they are entering before
hitting "submit". Probably will move logic to commit transactions to the submit
function, so that all updates are either committed or not.