Continuing to implement new front and back end for databases.
ZODB database is working, but need to more carefully read up on the locking and
commit mechanisms in order not to throw errors during normal program execution.

GUI frontend is starting to take shape but still need to finish implementing all
required functionality. May want to work towards having multi-paneled displays
for adding/modifying records eventually as well.

One issue that needs to be tackled is how to deal with whether a record or a file
is selected in terms of the 'modify' and 'remove' buttons. Firstly, need to sort
out how to reference back to the actual objects represented by the labels, and
then how best to change them.

Finally, for the add file dialog, need to improve the interface so that it allows
for OS-like file browsing to select. Might also want to add a utility in the
future that allows for selecting of multiple files and then subsequently chosing
whether to integrate them into an existing record or to add a new record with
the file chosen as an object?