Progress on database implementation.
Quick commit before potentially making substantial changes. Finally got basics of
modifying records to work, but realized that display and modification of files
should be changed. Plan is to try and implement a combobox style widget instead of
the normal entry, and to also add a file selector for choosing new files.