Fixed bug regarding database.
Realized that closing and re-opening database GUI caused issues as previous connection
to the database was not closed prior to calling open again. No two processes can attempt
to open the database concurrently. This may be something to keep in mind in the future
when considering upgrading to using multiprocessing. Either all database connections need
to be limited to single processes or need to try something like ZEO or RelStorage.