Beginning work on GUI frontend for searches.
Playing around with GUI frontend for searches. To begin, focus on how to setup a
single search from scratch, but eventually need to determine how to deal with
multiple searches, forward/backward searches, and getting results/summaries of
one or more searches. Most of this can be based off the initial implementation,
and will rely on how the database backend changes, similar to the database code.