Started interface for setting up queries.
Putting basic framework in place. Next step is to actually add queries from file.
Should probably add something about the query/file datatype (e.g. 'protein' or
'genomic'), as this is required, along with the record, for dealing with redundant
accessions. Also need to keep in mind that it might eventually be good to have a
feature to automatically detect redundant accessions, which could then be applied
to a number of queries in a non-blocking call to a hidden process/thread. Next step
though is to work on the backend database. Should decide whether each separate
element could be attached directly to root or whether they should be encapsulated
in data structures below root.