First crack at adding backend database for searches.
Realized that perhaps simplest way of dealing with databases is to uncouple
object creation/manipulation from storage. That way, the same code for the
actual database can be used and the code interfacing with specific parts of
it can be responsible for knowing what the objects being stored/retrieved
should look like and what methods can be called on them. If this works for
searches, will go back and change the database code to reflect this.