Testing new database structure.
Changing to central database for all central aspects of Goat. Initial code
appears to work but need to test whether added records persist within the
OOBTree object properly. Also need to consider that these structures require
total ordering of keys (they either need to be immutable, or they need to
ensure they are not changed after creation), so there has to be ways of
ensuring that this is maintained over time even with changes to objects
after creation.