Starting to sort out queries.
Decided that it would be beneficial to have queries as another header under the
main menu, so that users can manage queries outside of the guise of performing
searches (both queries and databases should still be able to be modified within
the search menu as well though).

Planning to model the interface loosely off of the one already in place for
records, with ways to add, modify, and remove both queries and sets of queries.