Continuing work on queries.
Laying out basic design for windows to add queries. Currently employing windows
to list all possible queries, those that have been added already, and information
regarding these queries. Need to consider a few things though:

1) Can keep the information window, but it complicates selection within the listbox,
as currently have it set to 'extended' to allow multiple selection of items. This
makes it easier to add groups of queries all at once but might confuse the display?

2) When adding by file, how do we name individual queries? We can assign a default
name for them, but this is something that would then have to be changed after
addition as opposed to adding manually, where it is possible to add the name directly.

3) When adding RAccs, especially in the case of adding by file, does it make sense
to pop up a separate window for each instance of "manual" choice, or simply run the
BLAST search and allow the user to deal with each in turn?