Almost finished with initial layout.
Starting first with initial layout of everything before attempting to hook up pieces
and add functionality. Note, decided on implementations for two things:

1) Removed middle frame from initial query adding window, but might add an option
to view and update information later. Either way, this can be changed in the main
query overview window so it is not necessarily a big issue.

2) For similar reasons, decided not to worry about names for each individual sequence
when adding queries by file (can be changed later if desired)

Next need to outline the interface for redundant accessions (expecting it to be
either similar or identical to the current way of adding queries). After this is
complete, can begin to add functionality.