Added GUI for redundant accessions.
Code for appearance is currently essentially identical to that for adding queries,
i.e. two listboxes with linked function; this may change in future depending on
how the interface works in practice.

Next need to add functionality. Probably easiest way to start is to try to add
queries using both methods, and then to deal with aspects that pertain to their
maintenance after having been added (this will also allow testing of the DB and
whether or not the current implementation of sets/groups works as expected).