Continuing on with GUI implementation for all aspects of searches.
Working now on a GUI implementation for adding and managing queries. Several
aspects to keep in mind while working on this:

1) Need to be able to add queries from files and also via manual input for single
queries at a time.

2) If adding redundant accessions is required, need to ensure that a record is
specified for the self vs self BLAST. Also, need to sort out how to run this in
the background as a non-blocking operation to allow the user to still perform
actions.

3) Want to implement interfaces not only for individual queries but also for
query sets (groups of queries) to make it easier to perform repeated searches
with the same queries.

Current commit represents the first attempt at all of these aspects. Need to
test and refine the current implementation.