Trying to work out a way to conduct self-BLAST searches in non-blocking fashion.
Continuing to work on query interface. Can add queries by file without apparent
issue, but currently cannot work out a thread-based implementation for running
needed self-vs-self search operations in the background. Goal is to spawn a
non-modal dialog box that monitors progress of all searches while these are
carried out by threads; upon completion of each search, the threads signal to
move the progress bar forward and then once all are finished the main GUI is
signaled back to in order to update the listbox elements.

Most of the implementation is in place but cannot trace a bug regarding the
actual searches within each thread. Threads should have the target of each
query instance's self_blast method, but calls to return the record_db object
for the instance's associated self.record variable fails because this is
somehow replaced by the integer '0', which could also mean a None value? It
is possible that the thread is using a newly minted instance with all default
variables, but this requires investigation to determine if this is the case
and also why the strange behaviour is present in the first place.