Working on modifying added queries.
Working on modifying query information. Currently, redundant accessions are a
combination of title and evalue; storing only that information rather than
whole BLAST objects in case this becomes an issue for storage later on. Next,
need to sort out shifting those raccs from one column to the next (similar to
adding query interface) and also picking up any changed data in the info frame
on "Submit" button presses. Should not need to call any updates on list or tree
views in the main window, since these just hold weak references to the object
in the underlying database (unlike add/remove calls).

One thing to sort out eventually is to have the title of each racc hit change
with screen redraws so that the evalue is always visible. This is a general
issue to deal with at some point as it also affects the way that info is
displayed in Label-based widgets such as the database and query information.