Fixed listbox item addition/removal in place.
For adding queries, it is not important to keep the listbox items in order
(although might change to do this anyway), but for Raccs it is critical to
maintain the order regardless of where they are in the list. Therefore,
updated the add and remove functions in the gui_util superclass to allow
for addition either by index or default as before (each item is added to
the end of the list), and fixed removal so that items are removed
accurately by index in the face of a shrinking list.