Modifying queries now works.
Modifying single queries now works. Redundant accessions can be chosen and modified
along with other query attributes. May eventually change which attributes can be
changed/how the interface is structured, but for now will leave it.

Next step is to deal with query sets. This may or may not serve as a model for DB
sets as well, with the idea that one could search the same set of queries across
multiple DBs easily.