Linked query set GUI to main query GUI.
Adding and modifying sets appears to work as expected, and changes are communicated
back to the main tree view widget. Of note, currently changes are first applied to
the query_db and then subsequently the whole tree is re-drawn from scratch. This
might not be ideal in the long-run; some ideas to improve this in future:

1) Totally decouple changes in the GUI from changes in the database until the user
hits 'submit'; this would require keeping all changes in memory until they are
made to the DB and committed, or at least holding all commit() calls until this
point - the latter might make more sense in the long run.

2) Improve the update() function - new sets are added to the end, and sets that
modify in place simply remove the relevant item and replace it in frame; nodes
in the tree that remain unchanged are not modified at all.

Next need to add a simple interface for removing query sets. After this, will go
back and work on query-based operations again, as queries can also be modified/
removed from the treeview, and changes (especially removal!) need to reflect in
all places where the query(ies) exist.