Finished initial draft of adding databases to searches.
Databases are currently added by record. This does have the unfortunate side effect
that it is possible for the user to add a record without the associated database that
is the target of the search (i.e. no protein file for a search targeting protein).

For now, assume the best case scenario and continue working on making searches
actually work. Can come back later and deal with improving the search setup and
interface. Next step is to deal with creating the search object and actually
running the search.