Added an implementation for intermediate queries.
Current implementation involves a sub-node for searches in the original QueryDB,
which is populated with nested object instances of the same generic storage class.
Each search can be retrieved and then iterated through to retrieve objects by
id, which then have their own trees of query objects. By keeping the qids in a
list in the result object (in the ResultDB), order can be maintained while still
having objects dispersed in an OOBTree.

Depending on how this works in terms of actually organizing and running searches,
this implementation may change in future.