Finished code to run forward and reverse searches.
Seems like nesting queries in results in searches within the QueryDB might not be
the best way to do things? Likely better for organization, but because searches
are run with a flat list of queries, requires a lot of looping to find the right
query; reducing it to one level of nesting (adding all queries directly to search
object) might improve lookup if there are a lot of queries to go through.

Either way, set up and test a proper search now.