Working on perfecting the reverse hit criteria logic.
Program appears to work, at least in terms of filtering forward/reverse results
and going through hits. Need to run more tests with print statements to ensure
that:

1) There are no unwanted/incorrect comparisons being made
2) The number of spurious comparisons are being limited

Then need to work on the evalue cutoff comparisons. For 'next_evalue' what is
needed is not a hard value but rather an orders of magnitude cutoff between
logarithmic values, which should be feasible.

Lastly, realized that redundant_accs added to the query object have both the
title and evalue; though not technically a problem, may want to fix this at
some point in time.