Fixed next hit evalue cutoff implementation.
Next hit evalue cutoff is an order of magnitude threshold, therefore needed to
convert the raw evalues to log(10) values for the comparison. Absolute value
ensures always comparing positives, and the order of hits determines whether the
value should be less than or greater than the specified cutoff. Could eventually
allow cutoffs for pos/neg and neg/pos (in terms of first hit) to be different but
for not keep things relatively simple.

Next need to sort out how to deal with 0.0 value evalues (instead of just making
them an arbitrary value), and trying to determine if there is a way to make these
comparisons without so much looping.

Finally, once this is done, can add the summary to the SummaryDB and work on, at
the very least, getting spreadsheet output of the final results.