Tried to fix E-value 0.0 issue without success.
Will have to open a bug report for this at some point in time. Issue is that BLAST
stops reporting E-values past a certain point (supposedly 1e-179) and defaults to
0.0 instead. This makes E-value comparison and use of magnitude cutoff criteria
difficult. Tried to calculate the actual E-value of these hits using either the
Score or the Bit-score, but standard formulas did not yield the same value as that
which BLAST reports (for hits with an actual reported E-value).

Unsure of why this might be the case, and unfortunately comparison of "E-values" for
two hsps, each calculated for the "wrong" value, did not yield the same order of
magnitude difference as comparison of the actual E-values, suggesting that the
calculations are not consistently off by the same margin.

For now, changed E-value re-assignment from 1e-300 to 1e-179 as the most reasonable
"placeholder" value, but will try to return to this issue at some point in the
future.

Next will work further on the comparisons to see if looping can be reduced, and
then work on an interface for viewing and exporting summaries.