Continuing improvements to searches/summaries.
Fixed issue in search_runner that prevented adding entries to int_queries, and also
added name of search to result objects to ensure that they will be unique. This
might change later, but for now seems to be the best way to prevent same query/db
pairs from breaking treeview display. Comparison of both rev_query.original_query
to fwd_search.query and fwd_uid.int_queries to rev_uid appears to prevent some of
the unnecessary looping. Would be nice also to eventually implement some kind of
function call to search a set of reverse results, which could then be called with
a shorter list each time one is identified as a matching result in order to
progressively reduce search space, although this would be hard to manage if we
wanted to parallelize this code in future.