Working on displaying information for summary objects.
Working on displaying informatin for summaries through the usual nested treeview/
label display framework used before. Managed to hit upon an interesting point
about treeview, which is that using the actual objects themselves as ids prevents
clash for same-named items as attributes of different objects (for example, this
solves the issue with results without having to add the search name to the result
identity; can just use the object itself as id.

Additionally, found that treeview ids can be associated with parents as well,
simplifying lookup from database for nested objects. Perhaps this could be used
as well within other nested objects outside of the context of treeview in order
to simplify lookup? E.g. for intermediate queries when setting up a reverse
search?

Need to finish up information display for other objects within summaries and
then work on getting tabular output.