Continuing work on displaying summary information.
Almost finished, but for some reason, hits are currently not displaying any
information, and in fact, it appears as though the retrieved object does not
have any attributes?

Additionally, there is still an issue with treeview object IDs. Now believe
that the id is irrelevant so long as the text label can be associated with
the object, either on its own or through its parents/children. Therefore,
considering trying to use a counter class to generate unique ids for every
single item in the tree. This would allow repeated objects, and also
provide a mechanism to retrieve previous IDs for inserting new objects by
retrieving the current value of the counter.

Will try this first, and then move on to working out why the hits are not
being properly stored/displayed.