Hit information is collected and retrieved correctly.
At this point, display is still compromised because of length of the hit titles.
Will need to revisit this at some point to change text length depending on the
size of the widget.

Next, beginning work on outputting tabular summary information. Will create a
rough draft from scratch and then later attempt to convert to using built-in
csv writing/reading utilies, if possible.