Fixed a small bug preventing DB status from updating properly when summarizing.
Changed from checking the state of the status variable to checking for a private
Boolean that is set to True the first time the status is determined. Issue likely
occurred because status was otherwise set to negative on first negative fwd hit
and then the conditional was not checked on subsequent hits. Could still cause
issues, in cases where status is set to tentative or unlikely first and then a
positive hit comes up later? Further tests will determine whether this needs to
be revisited.