Working on adding tabular output for summaries.
Added a small GUI interface for users to choose information about the summary to
output and linked this to a class for writing tabular output. Want to see if there
is a more robust built-in tool, but for now this should suffice. Appears to work,
but need to sterilize each entry to remove commas before writing.