Quick change to use IDCounter in building the result tree.
Turns out even when the same result is run multiple times, the old way of dealing
with inserting items could still throw an error on 'duplicates'; using the counter
solves this issue.

Should eventually make it so that DB is checked first for duplicate entries (at
least names) and asks user to choose a different one.