Finished first draft of threaded search implementation.
Fleshed out widget with a Static label above the progress bar and a dynamic label
under the progress bar. For some reason, tkinter variables were not working for
any of the elements in the window, whether be the progress bar or the text labels
and so had to resort to setting these manually each time they are supposed to
change. Unsure of why this occurred, but should be fine.

Next need to look into why files are not being deleted after searches when the
'keep_output' flag is set to True... could be an issue with setting the flag on
the object, or to do with the actual removal code itself.

Finally want to revisit settings/database/main start up code. For settings, could
stand to re-do the GUI interface and to also improve access to/changes of main
file. Some settings might be better off as private? (Or at least hidden from the
user?). For databases, considering a way to deal with sharing objects across
modules, such as using a separate module that could then be imported. This would
alleviate the need for closing/re-opening databases for each separate window (and
in fact is likely a necessity if the user ever wishes to have more than one window
open at a time that requires access to the same database). Then, exits in other
windows (not root) would simply either commit changes or not, whereas closing the
main window would actually close the DBs. This would require though the use of
some kind of global mechanism to prevent concurrent database writes (are reads
safe this way or is this also an issue?).