Fixed issue about keeping output files.
As expected, small issue caused output files to always be kept. Compared value to
the tkinter variable itself instead of the value it was set to so comparison always
hit else clause and set flag to True.

Also, fixed a minor issue with the search bar so that it starts the display at 1
instead of 0, i.e. it counts up to the proper number of searches (which is the length
of the list and therefore not zero-indexed)