Finally tackling some systemic issues with Goat.
First issue is that database connections are not global objects and had to be
passed down through widget heirarchies in order to be accessed. To fix this,
have database objects be populated on program instantiation and then imported
as global variables through a dict container in the initialization module.

Changed code for initializing Goat and tested this import with the Record DB
GUI; seems to work. Will continue to test this assumption with the other GUI
interfaces to database components to ensure this is ok. Then, need to change
all code for closing or submitting windows to only either pass or commit changes,
but not to close the file storage objects any longer - this should only be done
on main window exits!!!