Integrating new global database model using a single module and imports.
Initial testing seems to indicate that this works, and it allows to have multiple
windows open at the same time without breaking because window open/close events
do not have any effect on the status of the underlying database (unlike before).

Question remains whether or not there is a need for a global locking mechanism?
As long as all database update events are confined to the main GUI thread, it
should theoretically be possible to escape this need, but there is the case
where long-running operations in a separate thread may attempt DB access on
callbacks when they finish at the same moment that the user attempts to either
read to or write from the database.

One possibility is the use of a single global DB lock? This could simply be a
lock in the datbase file itself that must be acquired each time either a write
or read command is issued. This would also fit nicely with the model, i.e. that
each 'separate' database is actually just a wrapper for a node on the underlying
database.